+++ src/app/demo/components-overview/autocomplete-example/autocomplete-example.component.ts
@@ -3,6 +3,7 @@ import { FormBuilder, FormGroup, ValidatorFn, Validators } from '@angular/forms'
 import { RenderPropertyItem } from './render-property-item';
 import {
   exampleErrorCallback,
+  examplePickValueFn,
   logResult,
   setRequiredValidatorForFormControl
 } from '../../example-base/example-base-util/example-base-helper';
@@ -19,10 +20,10 @@ export class AutocompleteExampleComponent implements OnInit {
   showOutputEvents: boolean = false;
 
   options = [
-    { label: 'Option #1', short: 'O1', value: '#1' },
-    { label: 'Option #2', short: 'O2', value: '#2' },
-    { label: 'Option #3', short: 'O3', value: '#3' },
-    { label: 'Option #4', short: 'O4', value: '#4' }
+    { label: 'Meine Aufgaben', short: 'MA', value: 'A' },
+    { label: 'Gruppenaufgaben', short: 'GA', value: 'B' },
+    { label: 'Zurückgestellte Aufgaben', short: 'ZA', value: 'C' },
+    { label: 'Vertretungsaufgaben', short: 'VA', value: 'D' }
   ];
 
   renderProperties: RenderPropertyItem[] = [
@@ -61,6 +62,8 @@ export class AutocompleteExampleComponent implements OnInit {
   errorMessage: string = 'Das Feld enthält keinen gültigen Wert';
   errorCallback = exampleErrorCallback;
   errorCallbackString = this.errorCallback + '';
+  usePickValueFn: boolean = false;
+  pickValueFn = examplePickValueFn;
 
   // endregion